requirements.txt
# Core Web Framework
Flask==3.1.0

# Text Processing & NLP
spacy==3.8.7

# Document Processing
python-docx==1.2.0

# PDF Processing
PyPDF2==3.0.1
pdfplumber==0.11.7
reportlab==4.2.5

# Fake Data Generation
Faker==37.6.0

# LangGraph & LangChain (Optional - for advanced workflows)
langgraph==0.6.3
langchain==0.3.27
langchain-community==0.3.27

# HTTP Client (for LLM calls)
httpx==0.28.1

# Environment Variables
python-dotenv==1.1.1
